+++ src/pages/User/index.vue
@@ -1,7 +1,6 @@
 <template>
   <user-layout
     v-model:new-password="newPassword"
-    v-model:new-avatar="newAvatar"
     :user="user"
     :is-form-loading="isLoading"
     :see-password="seePassword"
@@ -9,7 +8,6 @@
     @see-password="seePassword = $event"
     @edit-password="isEditingPassword = $event"
     @save-password="(newPassword) => onUpdateUserPassword(user.id, newPassword)"
-    @save-avatar="(newAvatar) => onUpdateUserAvatar(user.id, newAvatar)"
   />
 </template>
 
@@ -19,7 +17,6 @@ import { ref } from 'vue';
 import UserLayout from './User.layout.vue';
 import { getUser } from 'src/utils';
 import { handleUserPassword } from './utils/handleUserPassword';
-import { handleUserAvatar } from './utils/handleUserAvatar';
 
 const user = ref<User>(getUser());
 
@@ -30,6 +27,4 @@ const {
   isLoading,
   onUpdateUserPassword,
 } = handleUserPassword();
-
-const { newAvatar, isLoadingAvatar, onUpdateUserAvatar } = handleUserAvatar();
 </script>
